Completing CRUD on reviews

+++ src/main/java/com/example/SoundVinyl/domain/model/Album.java
@@ -33,7 +33,7 @@ public class Album {
     private Double ratingAvg;
 
     @Column(name = "rating_count")
-    private Integer ratingCount;
+    private Long ratingCount;
 
 
     @Column(unique=true)
